Update ha.py
In the synchronous mode cluster, sync Node and other follow node's role are all replica. If  want to mount a Slave_VIP to Sync Node, it's not so convenient.Here, change the role of Sync Node to Sync_standby. When the cluster synchronization node appears, change the role of Sync Node from replica to Sync_standby. Similarly, when handover occurs, it will be degraded to replica